eunlg/requirements.txt
appdirs==1.4.3
aspy.yaml==1.3.0
attrs==19.3.0
black==19.10b0
boto3==1.11.3
botocore==1.14.3
bottle==0.12.17
certifi==2019.11.28
cfgv==2.0.1
chardet==3.0.4
Click==7.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
greenlet==0.4.15
identify==1.4.10
idna==2.8
importlib-metadata==1.4.0
importlib-resources==1.0.2
isort==4.3.21
jmespath==0.9.4
joblib==0.14.1
mccabe==0.6.1
meinheld==1.0.1
more-itertools==8.1.0
mypy==0.761
mypy-extensions==0.4.3
nodeenv==1.3.4
numpy==1.18.2
pandas==1.0.3
pathspec==0.7.0
Pillow==7.0.0
pkg-resources==0.0.0
pre-commit==1.21.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.22.0
s3transfer==0.3.0
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
toml==0.10.0
torch==1.3.1+cpu
torchvision==0.4.2+cpu
tqdm==4.41.1
transformers==2.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.7
virtualenv==16.7.9
zipp==2.0.1
